backend/med_bot3/requirements.txt
# Requirements for chatbot.py and rag_pipeline.py: includes pytesseract (OCR), docx2txt (Word docs), Pillow (images), transformers/torch (BLIP), numpy, tiktoken, supabase (pgvector), fitz (PyMuPDF), pydantic, and all other dependencies
fastapi
uvicorn
pymupdf==1.23.21
supabase
psycopg2-binary
requests
tiktoken
pytest
requests_mock
google-generativeai
Pillow
arcadepy
PyMuPDF
reportlab
pytesseract
docx2txt
transformers
torch
torchvision
torchaudio
torchtext
numpy
pydantic
youtube-transcript-api
SpeechRecognition
pyaudio
anthropic>=0.25.0
playsound
PyPDF2
flask
jinja2
python-multipart
sentence-transformers
tqdm
python-dotenv
